Code from AI runs
I would like to create a egui rust based display interface to allow me to display a rouge like game.  There will be a central panel that displays the world view.  We will work on the world view later.  It will need a panel to the side for information.

Lets update it to use eframe 0.31 and egui 0.31